Extract components from batch details
